… tests/Benchmarks/System.Buffers/Reader_Construction.cs
… tests/Benchmarks/System.Text.Http.Parser/HttpParser.cs
… tests/System.Text.Http.Tests/HttpParserBasicTests.cs
… tests/System.Text.Http.Tests/HttpParserTests.cs