- Optimize Header parser BufferReader usage
- Optimize reader TryReadToAny for two items
- Optimize reader TryRead<T>

Goal here is to make sure the reader is working well for the parser.

**Parser before**
```
      Method |      Mean |     Error |    StdDev |
------------ |----------:|----------:|----------:|
 RequestLine |  69.73 ns | 0.3118 ns | 0.2917 ns |
     Headers | 147.80 ns | 0.6158 ns | 0.5459 ns |
 FullRequest | 230.65 ns | 0.6696 ns | 0.6264 ns |
```

**Parser after (no new reader aggressive inlining)**
```
      Method |      Mean |     Error |    StdDev |
------------ |----------:|----------:|----------:|
 RequestLine |  66.01 ns | 0.2759 ns | 0.2581 ns |
     Headers | 151.10 ns | 0.6065 ns | 0.5673 ns |
 FullRequest | 227.00 ns | 0.9968 ns | 0.9324 ns |
```

Slightly slower reading just the headers, but _much_ easier to follow. If the reader's `TryReadToAny` is marked as `AggressiveInlining` it would end up notably faster than it started:

**Parser after (reader TryRead AggressiveInlining)**
```
      Method |      Mean |     Error |    StdDev |
------------ |----------:|----------:|----------:|
 RequestLine |  66.05 ns | 0.4254 ns | 0.3979 ns |
     Headers | 137.62 ns | 0.9707 ns | 0.9080 ns |
 FullRequest | 219.71 ns | 0.6370 ns | 0.5958 ns |
```

At a 10% impact it seems like we should keep the `AggressiveInlining`.

For the binary reader it is another 17% faster than it was:

**Before**
```
                 Method |      Mean |    Error |   StdDev |
----------------------- |----------:|---------:|---------:|
              ReadInt32 |  74.46 us | 1.462 us | 2.746 us |
    ReadInt32_BigEndian | 121.07 us | 1.324 us | 1.238 us |
 ReadInt32_LittleEndian |  89.43 us | 1.756 us | 2.221 us |
```

**After**
```
                 Method |      Mean |     Error |    StdDev |
----------------------- |----------:|----------:|----------:|
              ReadInt32 |  63.58 us | 1.2667 us | 2.0812 us |
    ReadInt32_BigEndian | 112.13 us | 0.7395 us | 0.6917 us |
 ReadInt32_LittleEndian |  75.97 us | 1.4968 us | 1.4001 us |
```