@JeremyKuhne 2018-10-14
Scraped a bit more perf out- updated the comments above.