@KrzysztofCwalina 2018-10-15
The parser was heavily optimized in the context of TechEmpower, and was very sensitive to very small changes. @davidfowl, @pakrym, do you think we should merge these changes?