@pakrym 2018-10-15
Would be nice to try it with Kestrel and see how it affects plaintext performance.