@JeremyKuhne 2018-10-15
> The parser was heavily optimized in the context of TechEmpower, and was very sensitive to very small changes.

If we need more perf tests I can add them. There is no way we can make a safe parser as fast as an unsafe one, of course. We can, however, try to make the fastest safe parser possible. Whether I do that with these existing classes or fork another set of classes is something I need guidance on- I don't know what the expectations are here.

Most of what I won here was through smarter logic (at least in the case of the perf tests we have). As stated, it would be even faster if I flipped it all to raw pointers and inlined code. There is just no way around that. 

> Would be nice to try it with Kestrel and see how it affects plaintext performance.

How do I do this?