@JeremyKuhne 2018-10-26
@pakrym, @davidfowl, @stephentoub, @ahsonkhan 

I've updated the pull to use the reader for the request line and also improve the construction of the reader. Getting the first span out of the `ReadOnlySequence` is extremely expensive (like 15-20ns). I've temporarily leveraged the intended API additions that @ahson and I discussed from investigating this.

This essentially gets rid of unsafe, better validates CR/LF pairs, and improves most of the parsing. Parsing the request line itself is slightly slower (5%, but is more correct in validating CR/LF). Reading the rest of the header is significantly faster ( 30%). The full request is about 15% faster- it would be faster still if we would use the same reader to parse the full request.

Here is where I started (updated for the latest bits in master:

```
      Method |      Mean |     Error |    StdDev |
------------ |----------:|----------:|----------:|
 RequestLine |  66.18 ns | 0.4604 ns | 0.4306 ns |
     Headers | 160.09 ns | 1.6956 ns | 1.5861 ns |
 FullRequest | 244.85 ns | 2.6407 ns | 2.4701 ns |

                 Method |     Mean |     Error |   StdDev |   Median |
----------------------- |---------:|----------:|---------:|---------:|
 ConstructSingleSegment | 30.71 ns | 0.6348 ns | 1.177 ns | 31.31 ns |
  ConstructMultiSegment | 25.54 ns | 0.5252 ns | 1.012 ns | 26.17 ns |
```

**After:**

```
      Method |      Mean |     Error |    StdDev |
------------ |----------:|----------:|----------:|
 RequestLine |  69.54 ns | 0.3191 ns | 0.2829 ns |
     Headers | 121.82 ns | 0.9522 ns | 0.8441 ns |
 FullRequest | 211.13 ns | 0.8298 ns | 0.7356 ns |

                 Method |     Mean |     Error |    StdDev |   Median |
----------------------- |---------:|----------:|----------:|---------:|
 ConstructSingleSegment | 14.14 ns | 0.3037 ns | 0.3376 ns | 14.22 ns |
  ConstructMultiSegment | 19.02 ns | 0.4060 ns | 0.8293 ns | 19.50 ns |
```

**IF** we were to create an unsafe reader constructor we could squeeze out another 3 ns each time we construct it. Bad user code that tears the structs in the `ReadOnlySequence` would cause us to violate type safety instead of throwing invalid cast exceptions.

cc: @joshfree, @danmosemsft