@JeremyKuhne 2018-10-26
Here is what passing the reader along instead of `ReadOnlySequence` gets you:

**Before**

```
      Method |      Mean |     Error |    StdDev |
 FullRequest | 244.85 ns | 2.6407 ns | 2.4701 ns |
```

**After**

```
      Method |      Mean |     Error |    StdDev |
------------ |----------:|----------:|----------:|
 FullRequest | 173.62 ns | 1.9351 ns | 1.8101 ns |
```

244.85 -> 173.62. Not too shabby.