@pakrym 2018-10-26
> passing the reader along 

What do you mean by this?