@JeremyKuhne 2018-10-26
>> passing the reader along

> What do you mean by this?

The current signature takes a `ReadOnlySequence<byte>`, here is how I changed it for the benchmark:

``` C#
[Benchmark]
public void FullRequest()
{
    BufferReader<byte> reader = new BufferReader<byte>(s_plaintextTechEmpowerRequestRos);
    s_parser.ParseRequestLine(request, ref reader);
    s_parser.ParseHeaders(request, ref reader);
}
```

> Were you able to add this to kestrel? Ping me if you need help, would be awesome to see how it behaves in out perf harness.

I've started on it, and I will ping you.  :)