@JeremyKuhne 2018-10-26
Made a last pass through to clean out the old code and tweak inlining. Here are the final results, which are a little bit better:

```
Class Handler

      Method |    Before |     After | Diff |
------------ |----------:|----------:|-----:|
 RequestLine |  77.27 ns |  67.31 ns | 0.87 |
     Headers | 175.35 ns | 136.59 ns | 0.78 |
 FullRequest | 266.24 ns | 181.06 ns | 0.68 |

Struct Handler

      Method |    Before |     After | Diff |
------------ |----------:|----------:|-----:|
 RequestLine |  66.54 ns |  54.49 ns | 0.82 |
     Headers | 151.08 ns | 110.16 ns | 0.73 |
 FullRequest | 237.48 ns | 157.77 ns | 0.66 |
```

Pretty happy with this, especially given that the code is much simpler and there is no `byte*`.