@mjsabby 2018-12-31
This is really interesting. Is the plan here to graduate this as a general utility parser? Or is this for benchmarking only?